# docs/swagger.yaml
definitions:
  model.BaseResp:
    properties:
      code:
        description: 结果码
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.CreateFieldInfo:
    properties:
      comment:
        description: 字段注释
        type: string
      isList:
        description: 是否为 列表
        type: boolean
      name:
        description: 字段名
        type: string
      sort:
        description: 排序位置
        type: integer
      structID:
        description: 对应结构体ID, 当 FType 为 FieldTypeStruct 时 有用
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
...
  model.CreateProjectInstanceReq:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 对应 定义 id
        type: integer
      firstLinkDefineID:
        description: 第一个 link 信息
        type: integer
      firstLinkInstanceID:
        type: integer
      generateLanguage:
        description: 实现工程所用语言及版本
        type: string
      name:
        description: 工程名
        type: string
      paramList:
        description: 参数信息
        items:
          $ref: '#/definitions/model.ProjectParamInstance'
        type: array
    type: object
  model.CreateProjectInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.CreateStructReq:
    properties:
      comment:
        description: 字段注释
        type: string
      fieldList:
        description: 字段信息
        items:
          $ref: '#/definitions/model.CreateFieldInfo'
        type: array
    type: object
  model.CreateStructResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.DBField:
    properties:
      comment:
        description: 字段注释
        type: string
      constData:
        items:
          type: integer
        type: array
      fieldID:
        description: 字段ID
        type: integer
      fieldType:
        description: 字段类型
        type: integer
      isConst:
        description: 如果是固定值, 那么记录
        type: boolean
      isList:
        description: 是否为 列表
        type: boolean
      mapKeyFieldID:
        description: 当fieldType 为 map 时
        type: integer
      mapValueFieldID:
        type: integer
      name:
        description: 字段名
        type: string
      parentID:
        description: 对应 父级 ID (可能还是字段, 也可能是结构体)
        type: integer
      sort:
        description: 排序位置
        type: integer
      structID:
        description: 对应结构体ID, 当 FType 为 FieldTypeStruct 时 有用
        type: integer
    type: object
  model.DBStructInfo:
    properties:
      comment:
        description: 备注信息
        type: string
      fieldList:
        description: 字段
        items:
          $ref: '#/definitions/model.DBField'
        type: array
      id:
        description: 结构体ID
        type: integer
    type: object
...
  model.DeleteProjectInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.DeleteStructResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的结构体ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.GetProjectInstanceListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 项目列表
        items:
          $ref: '#/definitions/model.ProjectInstanceInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetProjectInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.ProjectInstanceInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  model.GetStructListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 结构体列表
        items:
          $ref: '#/definitions/model.DBStructInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetStructResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.DBStructInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.ProjectInstanceInfo:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 对应 定义 id
        type: integer
      firstLinkDefineID:
        description: 第一个 link 信息
        type: integer
      firstLinkInstanceID:
        type: integer
      generateLanguage:
        description: 实现工程所用语言及版本
        type: string
      id:
        type: integer
      name:
        description: 工程名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.ProjectParamInstance'
        type: array
    type: object
...
  model.ProjectParamInstance:
    properties:
      comment:
        description: 字段注释
        type: string
      id:
        type: integer
      inputConst:
        description: 如果是来自确定值, 那么记录这个值
        items:
          type: integer
        type: array
      inputType:
        description: 入参来源
        type: integer
      inputVarDefineID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarInstanceID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarIsLink:
        description: 如果来自变量, 那么记录是来自 其它link还是其它Node
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      name:
        description: 字段名
        type: string
      outputConst:
        description: 如果是确定值, 那么记录这个值
        items:
          type: integer
        type: array
      outputType:
        description: 返回值类型 (如果返回变量, 那么不需要记录, 到时候直接用就可以)
        type: integer
      projectInstanceID:
        description: 所属节点id
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
...
  model.UpdateProjectInstanceReq:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 对应 定义 id
        type: integer
      firstLinkDefineID:
        description: 第一个 link 信息
        type: integer
      firstLinkInstanceID:
        type: integer
      generateLanguage:
        description: 实现工程所用语言及版本
        type: string
      id:
        type: integer
      name:
        description: 工程名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.ProjectParamInstance'
        type: array
    type: object
  model.UpdateProjectInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.UpdateStructReq:
    properties:
      comment:
        description: 备注信息
        type: string
      fieldList:
        description: 字段
        items:
          $ref: '#/definitions/model.DBField'
        type: array
      id:
        description: 结构体ID
        type: integer
    type: object
  model.UpdateStructResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的结构体ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
info:
  contact:
    email: guoyueyang@126.com
    name: author:guoyueyang
    url: http://www.amazingsaltedfish.com
  description: this is [amazing salted fish] swagger api page.
  license:
    name: License:MIT
    url: https://spdx.org/licenses/MIT.html
  termsOfService: http://www.amazingsaltedfish.com
  title: Amazing Salted Fish API
  version: 0.0.1
paths:
...
  /project/instance/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateProjectInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - project实体信息
      - 新建
      - 实体
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateProjectInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateProjectInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - project实体信息
      - 更新
      - 实体
  /project/instance/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteProjectInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - project实体信息
      - 删除
      - 实体
  /project/instance/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProjectInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - project实体信息
      - 获取
      - 实体
  /project/instance/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetProjectInstanceListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - project实体信息
      - 获取
      - 列表
      - 实体
  /struct/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: structInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateStructReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateStructResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - 结构体信息
      - 新建
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: structInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateStructReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateStructResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - 结构体信息
      - 更新
  /struct/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: struct的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteStructResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - 结构体信息
      - 删除
  /struct/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetStructResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - 结构体信息
      - 获取
  /struct/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetStructListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - 结构体信息
      - 获取
      - 列表
swagger: "2.0"